Rename certificates.
- What was previously called a "Heavyweight Delegation Certificate" will
  be callled just a "Delegation Certificate"
- What was called "Lightweight Delegation Certificate" will be called
  "Operational Key Certificate"

Only the former kind of certificate is actually used for delegation. In
the first iteration of the design that dates back to before the Byron
release, both kinds of certifiactes were planned to be used for
delegation, so the names made sense. With the current design, they are
rather prone to add confusion.

Also, this should help us avoid any confusion between the Byron
"Heavyweight Delegation Certificates" and the Shelley "Delegation
Certificates".