# This is a combination of 8 commits. # This is the 1st commit message:
Alter the endorsement registration rule, so that the set of confirmed proposals remains the same.

# This is the commit message #2:

Perform cleanup of confirmed proposals in  the epoch change rule.

# This is the commit message #3:

Remove the section that expressed that there were no voting deadlines. There are now.

# This is the commit message #4:

Define what "endorsement" means in this context.

# This is the commit message #5:

Explain the acronyms used in the labels of the update-proposal interface.

# This is the commit message #6:

Add a figure explaining the situation with a too small endorsement window.

# This is the commit message #7:

Model the software update proposal in the update proposal validity rule.

# This is the commit message #8:

Make it possible for an update proposal not to change the protocol version.