Add compatability for building with cabal
Switch package.yaml to -> <pkg>.cabal files. These are compatible with
both stack and cabal.

Add a top level cabal.project file to build all packages with cabal.

To use, e.g.:

  cabal new-configure -w ghc-8.6.2
  cabal new-build all
  cabal new-test all

Note that the Accounting module is missing from the accounting package.
Someone forgot to add that to the repo. So that package fails to build.