Fix doctest-discover use
Adding doctest-discover to the dependencies relies on a undocumented
feature/bug in cabal that brings all dependencies executables into
scope.  And also exposes all packages.  This does not work with
haskell.nix, which is more strict.

This also updates the file layout to repsect the new form that
stack-to-nix -o nix generates.