Slotting: Remove bodgy Enum-like functions
Originally `SlotId` had a `SlotCount` (now named `EpochSlots`) typeclass
constraint and was therefore able to have an `Enum` instance. When the
typeclass constraint was removed (because it was a pain in the neck), we
ended up with hand-rolled equivalents of `toEnum` and `fromEnum` methods
but they were never really used outside this one module, so let's drop
them.