[#385] Revamp ProtocolMagic/Id in Genesis datatypes
Instead of storing a @ProtocolMagic@ in @Genesis.Data@, and being
forced to pick a default for the omitted @RequiresNetworkMagic@ field,
switch to a @ProtocolMagicId@.

Add a @RequiresNetworkMagic@ field to @Genesis.Config@.

Leave @GenesisSpec@ the same - use its @ProtocolMagic@ to fill the
@ProtocolMagicId@ of @Genesis.Data@ and the @RequiresNetworkMagic@
of @Genesis.Config@.

This seems to simplify the JSON instances considerably. We will no
longer be backwards compatible with `cardano-sl`'s configuration,
but that is ok.