Support short-cutting for leader election / verification
This change allows to speed-up the following computation

`p < 1 - (1 - f)^sigma`

where `sigma` is the relative stake, f is fixed to 0.1 and `p` is a
probability.

Using the Taylor series expansion for the exponential function, one can estimate
the error remainder and stop at an early iteration if it is either clear that p
is less than or greater than the above.