Ensure resulting `appVersions` values are strict in `registerVote`
It is possible that `registerVote` can return an update interface
`State` which contains thunks within its `appVersion` field (a strict
`Map ApplicationName (NumSoftwareVersion, FlatSlotId)`). Because a
strict `Map` only ensures that keys and values are in WHNF, it is
possible for a thunk to exist within the `(NumSoftwareVersion,
FlatSlotId)` value. This exact thing seems to occur due to a list
comprehension within `registerVote` which utilizes a generator that can
produce lazy values.