Fix errors in the `UPIREG` generators.
This error get's triggered

    inMap an av (swVer <$> avs) ?! AVChangedInPVUpdate an av

but it shouldn't since it is only in the first rule, and the next rule should be tried.

With a maximum length of 1000, traces are generated around that size, however
the generator is really slow and takes a huge amount of memory. We should look
for lazy data structures in the update mechanism.

[skip ci]