Remove most Ord instances for crypto keys
The pattern that our crypto libraries use is that they deliberately do
not provide Ord instances for keys (verification or signing keys).
Instead our crypto library encourage the use of comparisons on key
hashes.

The one remaining Ord instance is for RedeemVerificationKey because the
AvvmBalances type uses it as a Data.Map key type, though it probably
should not.

Many tests do (quite reasonably) rely on Ord so this patch moves the
required orphan instances into the test suite.