Update the Overview of Delegation section for scripts
Clarify address vs value address. Add sections on key addresses vs
scripts. All at a high level, not too much detail.

I removed the detail about the payment key pair, since we don't need the
maths in the overview section, and if we kepy it, it would have to be
extended to cover the script case too, which is then far too much detail
for an overview.