Drop Aeson.To/FromJSON instances
They are not in fact used by any dependent library. We do not plan on
using them for the structured logging.

This also drops the round trip and golden tests for the Aeson instances.
All the canonical JSON instances and tests are preserved.

This also reduces the size of the compiled library by about 13%.