Remove StaticConfig
StaticConfig was a union of either an external genesis file or a genesis
spec. The idea was that you could construct a config from either of
these. This was a relic of a bad idea from the cardano-sl
implementation.

The bad idea was that a node could generate a genesis config on the fly
from a config file with the genesis specification. The idea was this
made it easier to run dev clusters.

It was always a confusing approach however. It meant that all the config
needed to make a genesis file needed to be included in the config used
for running the node. Nobody could keep track of which config items were
used directly by the node and which were only used to make a genesis
file.

It's much better to very clearly separate these stages: the node always
runs given a pre-generated genesis file; and separately there's a tool
to generate new genesis files for dev nets.