Move and rename mkConfig to generateGenesisConfig
Since it is now the only thing in the Config module doing generation we
can finally clean things up and have all the genesis generation stuff in
the Generate module, as something that sits on top of Config and not the
other way around.

So this is really the last step of abolishing the concept of generating
gensis config at node startup.

Genesis generation is just another feature that sits on top and is used
in the external genesis tool.