[skip ci] Use VRF algorithm from cardano-crypto-class
Abstract over VrF functions.

We have taken an alternative approach to the DSIGN and KES families, and not introduced wrappers around the VRF types. These wrappers are largely unnecessary and mostly require lots of (un-)wrapping. In the future I would like to remove them for DSIGN and KES as well.