Add new `DCert` type for moving intermediate rewards
Currently collects a mapping from reward addresses to values in the delegation
state. The mapping is updated in the case of a corresponding certificate which
carries a sequence of key hashes and values for the reward accounts.

Each key hash `hk` is used to construct a reward account which is accesible with
the corresponding signing key `sk`.