Remove support for HD addresses from genesis generation
Previously we could have either "plain" key addresses, or HD addresses
(though these only used the first derivation index).

The node and ledger should not need to know about HD addresses at all,
since it is transparent to the ledger and used only by wallets. This is
the only use of HD addresses in the library. Removing it will allow us
to remove a lot of other code.