Move the HDAddressPayload from crypto to ledger lib
Moved into the AddrAttributes module.

This is a prelude to removing the HD module from the crypto lib
entirely. This type is the only thing we still use. It is an attribute
that we treat specially in the ledger. We don't look at its contents at
all but we do exclude it from attribute length checks (as it is used on
the chain historically).