Design doc: no witness for stake key registration.
We have decided against requiring a witness for registering a stake key. The
only reason we had for requiring one was to prevent people from registering a
key they do not own (and they can still end up in the same situation if they
lose the key after registration). Not requiring a witness makes the reward
transfer from the incentivised testnet easier.