Remove unused withSafeSigner
The SafeSigner abstraction keeps a encrypted secret key which has an
assoicated scrypt hash of the secret key. This scrypt hash is used to
check that the decryption passprase for the secret key is correct.

This kind of functionality makes sense for a wallet but not a node.

The whole node and CLI tools compile with the withSafeSigner
constructor removed. The only other constructor is noPassSafeSigner
which, as the name suggests, does make use of the special feature of
SafeSigner.

From this we can conclude that nothing uses the special feature and it
could be simplified.