Change SafeSigner to use SigningKey not EncryptedSigningKey
These are sadly not well named. The SigningKey is in fact encrypted.
The EncryptedSigningKey wraps a SigningKey but adds the ability to
check that the passphrase used for encryption is the right one (by
keeping an scrypt hash of the unencrypted key to compare against).

But this passphrase feature of the EncryptedSigningKey is not used by
the SafeSigner so we can swap it over to the simpler SigningKey.