[#871] Add `RetirePool` generator and correctness property
In the process of implementing this I encountered a bug in hedghog's `Gen`
internals. We decided to not bother debugging this and to just switch the
generators and properties over to QuickCheck.

In this commit I duplicate the generator logic for QuickCheck, introducing a
number of new generator modules which have the suffix `.QuickCheck`. This allows
us to make a gradual transition over to QuickCheck, rather than forcing all
generators to be rewritten at once.

I also define shrinkers, and I use those & the generators in the `RetirePool`
property.

One notable design decision I've made is to define a generator and shrinker at
the type of `SourceSignalTarget` for POOL. This gives more minimal
counterexamples, since we are only concerned with the bits of the LEDGER trace
which are projected out to the POOL level. It also works around an issue I
encountered with the STS transition failing when I used `forAllTrace`.

Closes #871.