Allow for generation of MIR certs
Problem was that with `d = 0` no MIR is allowed. Therefore the valid signal test
failed and the shrinker tried to find a smaller CEX.

While doing this it removed elements from the Tx and arrived at empty Tx inputs,
because currently, but changed the reason for the invalid Tx.

The current solution is therefore to restrict d to non-zero.