Allow software updates to start from version 0 or 1
It does not really matter if application versions start from 0 or 1, and
the old cardano-csl code allows both.

Also update the comments for the function that does the validation of
the versions. The existing comments says that we allow keeping the same
version, but the existing code and the formal spec always require an
increment of exactly 1. So we correct the comment.