Follow the secret key generation pattern for AVVM keys
For some reason the AVVM keys were handled differently, in that in the
generateSecrets it did not actually generate the secret keys for the
AVVM addresses, but instead just generated the seeds that could later be
used to deterministically derive the secret keys.

This is unnecessarily complicated and does not followthe established
pattern for no good reason. So switch to actually generating the AVVM
secret keys in generateSecrets and then just using them later.

This will need some slight changes in the node CLI since it can now dump
the AVVM secret keys instead of the seeds, but it appears that these are
currently unused anyway.