Generate larger UTxO for more variance
Before, we generated at most 5 initial UTxO entries and created at most 2
outputs in a transaction which consumed betweed 1 and 5 inputs. Therefore the
size of the UTxO quickly converged to 1-2.

This commit changed this to generate between 10 and 100 initial UTxO entries and
then using somewhere between 1 and 10 as inputs. The number of receipients is
generated using then number of inputs, contracting by 1 or expanding by one with
a relative frequency of 1 and generating the same number of outputs as consumed
inputs, with a relative frequency of 2.