Added toCBORABlockOrBoundaryHdr and toCBORABlockOrBoundaryHdrSize
`ouroboros-consensus` is using
`Ouroboros.Consenus.Byron.Ledger.Serialization.encodeUnsizedHeader` for
that, but the right place for this encoder is in `cardano-ledger`, where
it is easier to write tests.

This patch only includes tests for encoded size estimations, a seprate
PR should be done to add a round trip test.  Then we can use it in
`ouroboros-consensus` in place of `encodedUnsizedHeader`.