Adapt code to use `ActiveSlotCoeff` type
This type is currently defined as the pair of the active slot coefficient value
and an integer that contains the logarithm of `(1 - f)` multiplied with the
fixed point precision. The second value should be the FixedPoint value itself,
but currently this does not work because of the missing `NoUnexpectedThunks`
instance of FixedPoint.