Fix EBB header's "body proof" hash.
In OBFT, the EBB body always contains an empty slot leader schedule. The
EBB header contains a hash of the body. That hash needs to match. It was
previously the hash of the empty CBOR bytestring (0x40) but it needs to
be the hash of the empty list, using the indefinite encoding (0x9F FF).