Expose the shelley-spec-ledger test helpers as a package
Similarly to how `cardano-ledger` exposes its test suite as the
`cardano-ledger-test` package, this commit exposes the generators and
examples (but not the tests themselves) from the testsuite of
`shelley-spec-ledger` as the `shelley-spec-ledger-test` package.

This makes it possible for consensus to reuse the existing generators and
examples for testing the Shelley ledger integration instead of having to write
them from scratch.