Introduce OBftSlot
To fix the thunk in the overlay schedule (map), replace `Maybe (GenKeyHash
crypto)` with a new isomorphic (up to strictness) type.

In consensus, we have to deal with the result of the lookup in this map, i.e.,
`Maybe (Maybe GenKeyHash crypto)`. As nested `Maybe`s can be confusing,
replacing the inner `Maybe` with a new type also helps with code clarity.