Introduce StrictMaybe and use it
Wherever we have `!(Maybe x)`, the bang won't forced `x`, just `Nothing` or
`Just _`, so we need a strict variant of `Maybe`. In consensus, we typically
introduce a new isomorphic type (up to strictness) for each such `Maybe`, but
there were simply too many, so I succumbed and introduced `StrictMaybe`.