Hashing: change the representation of Hash to ShortByteString
Change AbstractHash from being a newtype over Digest to a newtype over
ShortByteString and adjust the conversion functions.

The ShortByteString type has efficient Eq & Ord instances whereas the
Digest type has very inefficient ones.

Previously the Digest comparison loop was near the top of the profile
while doing a bulk sync. This representation change is an improvement
of ~7% in the time to validate the whole chain.