crypto: Deprecate hash function
The function 'hash :: ToCBOR a => a -> Hash a' serializes 'a' to CBOR
and then hashes it. Since a lot of data types are actually annotated,
performing the CBOR encoding is completely un-necessary, since the
existing annotation can be passed directly to 'hashRaw'.

We do however provide 'serializeCborHash :: ToCBOR a => a -> Hash a'
which is fine for tests and for the case where there is no annotation.