Staging exemption for a stray null update proposal
A "null" update proposal is one that neither increase the protocol
version nor the software version. Such update proposals are invalid
according to the Byron specification. However in the cardano-sl code
they are accepted onto the chain but without any state change.

We cannot follow the legacy cardano-sl interpretation of accepting null
update onto the chain with no effect because it opens the door to DoS
attacks by replaying null update proposals.

For further details see:

https://github.com/input-output-hk/cardano-ledger/issues/759
https://github.com/input-output-hk/cardano-ledger/pull/766

The existing staging network (protocol magic 633343913) does have a null
update proposal however, in epoch 44 (absolute slot number 969188). We
could delete the staging network blockchain and start from scratch,
however it is extremely useful for testing to have a realistic chain that
is as long as the mainnet chain, and indeed that has a large prefix that
was created by the legacy cardano-sl codebase. Therefore we allow for
this one specific excemption on this non-public testing network.