Call the 'TICK' rule in future ledger view.
We used to call the NEWEPOCH rule, but this was done with the
understanding that TICK changes related to the ledger view all happened
on the epoch boundary. This was true, but only because genesis
delegations took place inside the ledger rules, rather that in TICK.
This is problematic (it violates future time travel implementation), so
now genesis delegation updates happen in TICK and we call that.

Also, we used to not be able to call TICK because it was called BHEAD
and took the header. Now we can, so just calling it is easier!