Rewrite the Keys module and related code.
This turned out a larger change than I was hoping for!

We aim for two things in this PR:
- To remove, to the full extent possible, indirection over existing
crypto types in cardano-crypto-class via newtype wrappers introduced in
this package. As such, `SKey`, `*KeyES` etc are all removed, along with
the equivalents for hashing.
- However, we use DSIGN keys and hashes for multiple purposes over the
codebase. As such, we choose to retain the `VKey` wrapper because we use
it for differentiating on the _role_ that a key/hash is used for in a
given place. The full list of these is given in the `Keys` module. Just
from the experience of writing this PR, this makes interacting with
things that use keys much simpler, and helps catch bugs (one was caught
in the delegation infrastructure during this PR).

Whilst I believe this PR makes the library code much clearer, the
testing code alas uses the same keys all over the place, and hence is
now littered with `coerceKeyRole`. It would be nice to clear this up in
a subsequent PR.