Fix the examples.
This change should hopefully make the examples much less fragile. We now
pick the correct key based upon the slot, rather than relying on knowing
the order of nodes in the overlay schedule.

The new `slotKeys` function picks the correct set of keys for the given
slot. We keep the old `coreNodeKeys` only for cases where we are picking
keys for another reason. At some point I would like to get rid of this
altogether.