Make PRTCL block-independent
The PRTCL transition was doing block-specific checks:
* Hashes must align
* Slot number must increase
* Slot number must not be in the future (compared to wall clock slot)
* Block number must increase by one

These checks are *all* already done in consensus. Them being redundant is not
a problem, but in order to perform these checks, the PRTCL state stored
`LastAppliedBlock`. This coupled the *protocol* transition, which should be
independent/agnostic of the block, to the *block type* through its hash. One
could argue that this coupling should not be there.

The real problem is that when we hard fork from Byron to Shelley, we must be
able to translate the Byron protocol state to the Shelley protocol
state (called `ConsensusState` in consensus). At this point, we would have to
produce a `LastAppliedBlock` for the Shelley protocol state while the Byron
protocol state stores no such info. In consensus (and in the Byron ledger),
the protocol state is agnostic of the block, so we simply can't conjure up
such a `LastAppliedBlock`.

Moreover, having to know the current wall clock slot is also problematic,
because we don't always *know* the current wall clock slot. For example, when
we're validating blocks in the middle of the chain, there might be a hard fork
at a later point in the chain, which changes the slot length, so we simply
can't compute the current slot yet, because we don't yet know when the slot
lengths will change. In consensus, we have our own "not in the future" check,
which is implemented differently, making the check in the ledger redundant.
Actually, for this reason, the current Shelley integration in consensus just
passed the block's slot as the current wall clock slot, effectively making
this check useless in the ledger.

The PRTCL transition still needs to know the previous hash to pass it to the
UPDN transition, which uses this as extra entropy (converts it to a `Nonce`)
when a new epoch starts. Instead of storing and passing a `PrevHash`, we now
convert it earlier to a `Nonce`, and store and pass that `Nonce`. This
decouples the PRTCL (and UPDN) transition from the block type.

* All block-specific checks are removed from the PRTCL transition. The
  corresponding `PredicateFailure` constructors are also removed.

* The PRTCL state now stores a nonce derived from the previous hash instead of
  `WithOrigin (LastAppliedBlock c)` (previous block number + slot number +
  hash), as that info was only needed by the removed checks.

* The PRTCL environment no longer contains the current wall clock slot.

* The UPDN environment now contains a nonce derived from the previous
  hash (comes from the PRTCL state) instead of the previous hash. The only
  thing it did with the previous hash was to derive a nonce from it. Now we
  just do it earlier.

* The CHAIN state still stores the `LastAppliedBlock`. Consensus doesn't have
  an equivalent of the `ChainState`, so consensus doesn't care about this.
  Keeping that `LastAppliedBlock` there makes the test easier, i.e.,
  `mkBlock`.

* The CHAIN environment is now `()` instead of `SlotNo` (current wall clock
  slot number).