Move stack build to nightly checks.
Rationale: the stack build is the slowest part of CI. We're mostly using
it for generating coverage reports. But there's no particular need to
have these immediately. Moving to nightly ensures we still get to test
things build with stack, and we also have coverage, but hopefully faster
CI times as well.