Minor updates to the design spec sections on pool metadata
A few changes to bring the design spec in line with the actual current
design:

Move the IPs & DNS names from the pool metadata into the pool
registration certificate proper. We decided to move it there to make it
directly (rather than indirectly) accessible to all nodes to allow
using it to bootstrap the P2P network.

Trim the pool metadata fields down to the agreed ones.

Make the description of the stake pool metadata aggregation servers
better reflect the current design.