Separate hashing for addresses
- We now enhance the key role with a separate parameter indicating its
hash type. This allows us to distinguish between keys which are hashed
in different ways. We have two hash types - address, and regular.
- The `keyHash` function now performs the correct hash depending on the
type of the key.
- Witnessing now needs to distinguish between the various types of keys
used. So we now have various witness sets which contain both address and
regular witnesses.

These changes caused some problems for the tests, which freely mix key
roles all over the place. So we have also refactored the tests to be
much clearer about which keys are being used and how.

The `NonTraceProperties` are even worse than the normal tests in terms
of key usage, and here we give in temporarily and use `unsafeCoerce`,
which is safe in this case since we know that the hash types align. At a
lower priority, we should rewrite these tests to use `KeySpace`.

The `Address` test is currently disabled, since it did some strange
things mixing crypto types. We will open a ticket to rationalise and
re-enable it.

Fixing the key tests seemed to cause a breakage in one of our property
tests (registeredPoolIsAdded). On closer inspection, this test was
checking invalid properties, so it's a little worrying that it was
passing before. My best hypothesis is that the confusion about keys was
preventing pools from ever being _re-registered_ in this test, which
would have allowed it to pass. In any case, this was incorrect, so this
PR also includes a fix for that test.