Byron: replace tuples stored in maps with proper strict data types
The `NoUnexpectedThunks` check in consensus was failing for some of these.
Instead of sprinkling bangs and `force` through the code, solve it properly
once and for all by replacing the thunky tuples with new strict data types.

Important: the `From/ToCBOR` instances for the data types have to match those
for tuples to guarantee binary compatibility. Fortunately, when writing them
as you otherwise would, they do.