Avoid Double in JSON genesis parameter conversions
We use Rational internally for many parameters (either directly or via
UnitInterval). We must be careful with conversions of fractional numbers
in JSON.
In particular if we convert a value like 0.1 via Double which cannot be
represented precisely then we end up with this:

> toRational (0.1 :: Double)
3602879701896397 % 36028797018963968

Rather than 1 % 10

So instead we must go from the Scientific type that Aeson uses, directly
to Rational or (Ratio Word64 for UnitInterval) without going via Double.