`applyRUpd` now checks for key registration.
This commit fixes the underlying issue behind the missing ADA.

When we create a reward update, we only grant rewards to registered
stake keys. However, it's possible to deregister them after the update
is created but before it's applied. In such a case, `applyRUpd` was
adding them to a reward map anyway. When the key got re-registered, the
reward account was being reset to 0, and hence money was being
destroyed.

This commit adds a check to `applyRUpd` that verifies whether a key is
still registered and, if not, returns the funds to the reserves.