Add TICKN transition.
In order to produce a block, we need to create a VRF output using the
correct epoch nonce. This requires us to be able to calculate the epoch
nonce without providing a header. As such, we divide UPDN into two
rules, TICKN and UPDN. The former only does header independent updates,
whereas the latter is used to incorporate header data into the evolving
and candidate nonces.

The TICKN rule is then called directly from the CHAIN rule. In the real
implementation, TICKN will be exposed directly to the consensus layer.