fix a couple spec typos, github issue #1654
The formal spec did not remove the retiring pools
from the fPoolParams mapping in the POOLREAP rule.
(it was, however, correct in the implementation).
Additionally, in the spec and in the implementation,
there was a variable kh that should have been hk
to be consistent with the rest of the spec.