Minor changes to shelley.cddl for ease of use
In order for ease of code generation these changes, which don't impact
the CBOR representation, were made. These certificate/multisig changes
are the same ones that were done to relay before, as otherwise there is
not enough information for code generation tooling in each group choice
variant to get proper names from, as comments are discarded.

amount changing to coin was done to improve code generated and to make
it more clear for readers.

Signed-off-by: rooooooooob <raftias@gmail.com>