Check Byron address attributes are not too big
Or we'll have the same old problem as in the Byron era.

This is now a stricter limit 64 vs 128, and only applied for new txs
for their tx outputs, not in the witnesses. It does not affect any
addresses brought over from the Byron era.

We calculate the size based on both the variable-sized known attributes
and the variable-sized unknown attributes. You get a free pass on the
fixed-size known attrs.

The limit of 64 is adequate for the known existing uses by legacy
wallets, which should have attributes of around 32 bytes at most.