Start optimising the memory footprint of the UTxO.
This changes the memory representation of TxIn and TxOut, to reduce overhead.

In particular, we use unpacked Word64 for coin values and txout indices, and use
the serialisation of addresses as ShortByteString.

This does not yet change the representation of TxId.

The changes are nearly opaque, in that we provide view patterns that takes the
place of the old constructors. We do need to add a Crypto constraint wherever we
use a TxOut, however, since the view pattern relies on
serialisation/deserialisation.