Make the overlay schedule abstract
This will allow us to replace its implementation/representation by something
more compact and efficient.

On mainnet, the epoch size is 432,000, which means we're creating a map of
nearly half a million of entries just to implement a spaced round-robin scheme.

When writing ledger snapshots, we're also writing this map, albeit in a slightly
more compact representation (see `compactOverlaySchedule`), to disk.

Note that the size of this map will shrink when `d` goes down.

It should be enough to keep the arguments passed to `overlayScheduleHelper` in
memory and implement the current API of `OverlaySchedule` as a function instead
of a mapping. Similarly for serialising an `OverlaySchedule`, just de/serialise
these arguments.

To test this approach, the function-based implementation can be compared against
the current map-based implementation (e.g., by implementing
`overlayScheduleToMap`).