Hardwire bootstrap witnesses and key hashes to Ed25519
This patch allows the Shelley ledger to be instantiated with a DSIGN algorithm
other than Ed25519.

Previously, the Shelley ledger implementation was using the given DSIGN
algorithm for everything, including bootstrap keys and witnesses. The algorithm
had to be Ed25519 because that's the only way to validate a bootstrap (ie
Byron) signature.