Adjust how we define the script hash type, but keep it the same
This is preparation for changing the script hash type to use ADDRHASH
rather than HASH. We change a few things that depend on it being
specifically HASH.

Redefine ScriptHash directly in terms of HASH, rather than indirectly
in terms of the Hash type (which does also use HASH).

We have to drop the HashAnnotated instance since it only supports HASH
but we're about to switch to using ADDRHASH. We only use the instane in
one place, so we just inline it there.