Add a compile-time test ScriptHash and KeyHash matching
This compile-time test asserts that the script hash and key hash use
the same hash size and indeed hash function. We do this by checking we
can type-check the following code that converts between them by using
the hash casting function which changes what the hash is of, without
changing the hashing algorithm.

This test fails at this stage with a type error, as it should, but will
compile as of the next patch.