mkBlock no longer requires fake VRF
We were previously using `mkBlock` throughout our generator code.
However, 'mkBlock' was specialised to create blocks with fake VRF
output. As a consequence, we had a dependency on the fake VRF
throughtout tests, rather than simply in the examples, and we also had
to go through extra work to conjure up appropriate VRF values in our
generators.

This commit removes the ability to specify the VRF values from mkBlock
and the constraint on the VRF class in the `Mock` constraint. They are
reintroduced as `ExMock` and `mkBlockFakeVRF` for the use of the
examples only.

(OK, technically the 'NonTraceProperties' also use them, because I
haven't started to look at rationalising those. One thing at a time)