Reimplement 'genBlock'.
I managed to break it earlier through removing the fake VRF, but wanted
to be able to use it in a sensible way without the fake VRF. Also I
found it hard to understand. So this commit rewrites it. Benchmarks show
basically the same performance. Pushing to allow tests to run; if things
work well, I'll delete the old code.