Delegs: optimise the rewards calculation
From my profiling report of the validation of the first Shelley epoch (excluding the epoch transition):

```
 TOT   INH   IND
35.7  28.1    .4  Control.State.Transition.Extended liftSTS (7553)
13.0   9.8   7.8    Shelley.Spec.Ledger.STS.Delegs delegsTransition.rewards (8136)
 4.4   1.2    .6      Shelley.Spec.Ledger.Address >= (8607184)
 3.8    .6    .6        Shelley.Spec.Ledger.Address < (8607184)
 3.2     -     -          Shelley.Spec.Ledger.Credential < (8607184)
  .8    .8    .8      Shelley.Spec.Ledger.Address mkRwdAcnt (8615225)
11.3   9.0   8.6    Shelley.Spec.Ledger.STS.Delegs delegsTransition.rewards' (8136)
 2.7    .4    .3      Shelley.Spec.Ledger.Credential >= (8607184)
 2.3     -     -        Shelley.Spec.Ledger.Credential < (8607184)
```

The conversions (`Map.mapKeys`) of the rewards accounts is very costly, and
we're doing it twice. Instead, convert the `wdrls`, which is a much smaller map.

After this change, `rewards` and `rewards'` are no longer hotspots (0.0 time!).