Made some changes to the 'TxOut' pattern.
In particular the function 'viewCompactTxOut' which now tries to serialize the Address
directly from a ShortByteString, and falls back to the expensive 'deserialiseAddr' only
if that fails. Benchmarking shows that it almost always succeeds. This speeds up the
'applyBlockTransition' function about 10-12%. More speedup is possible
because the new bottle neck is the function 'substring :: ShortByteString -> Int -> Int -> ShortByteString'
which has an absolutely horrible implementation, it should use a single allocation and a memcopy
operation, but instead it goes through 2 intermediate lists.