Optimise DeserializeShort.substring
In my microbenchmark, this brings down the cost of deserialising a Shelley
address in `viewCompactTxOut` from ~360ns to ~60ns.

The `cloneByteArray` function was introduced in version 0.7.1.0 of the
`primitive` package, so update the lower bound accordingly.

I believe the old `primitive < 0.7` constraint in `cabal.project` was due to
older versions of the `cborg` package having `>=0.5 && <0.7.1.0` as bounds for
`primitive`, but `cborg-0.2.4.0` bumped the upper bound to 0.8.

This will likely require similar changes to downstream repos.