Implement API endpoint for pool sizes in terms of total stake.
This does not (yet) change the type, just the normalisation, for ease of
integration in the wallet backend in the next release.

In the long term, we will want to have a well-planned endpoint to deliver all
the pool data that the wallet might want (and have the format be extensible as
well).