Redo metadata decoding
Don't go via CBOR's Term, do it directly so we can enforce the schema
properly. For example the metadata CBOR schema does not support big
integer representations.

Also add a validation function for the things we do not enforce in the
decoder. This will need to be used in the tx validation rules.