Don't create a reward update in the future view.
The future ledger view carries out the TICK transition, which calls
RUPD. However, the result of the RUPD transition is thrown away
immediately. Since this is a potentially expensive computation, it's
nice to avoid.

This should address https://jira.iohk.io/browse/CAD-1897

This is implemented using a separate TICKF rule, outside of the normal
rule hierarchy. This is done to avoid changing the hierarchy, which
would result in a change to things which get serialised (the error
stack).