Add infrastructure for translations between eras
This introduces the `TranslateEra` type class (see its documentation) for
translating era-parameterised types from an era to the era after it.

The consensus layer will use this class and its instances to translate the final
Shelley (ledger/protocol) state to the initial Allegra state (same for
Allegra/Mary and subsequent eras). But also to translate some configuration
data, e.g., the Shelley genesis config to the Allegra genesis
config (`ShelleyGenesis era`). We currently intend to keep the Shelley and
Allegra genesis config representationally equal, yet the era tag mentioned in
the two types is different, requiring a translation.

We allow for passing per-era context to the translation as well as choosing the
error type per era/type combination. For example, hash translation could be
total, but transaction translation could be partial, but both can use the same
infrastructure.

The goal is to have a module for Allegra and a module for Mary with a bunch of
empty (using the defaults) and non-empty standalone instances of this class.