Get haskell-language-server working with c-l-s.
haskell-language-server is the project which merges haskell-ide-engine
and ghcide, and is now looking pretty good. Sufficiently good that I
think it's now plausible to use it to replace a ghcid workflow.

This commit adds various things to support using it in cls:
- Updates the index state to one which contains haskell-language-server.
- Adds a nix expression in haskell-extra.nix to build it. This is
completely cribbed from the Plutus codebase.
- Update cardano-prelude to include a fix for a change in
base16-bytestring.
- Add a script which runs haskell-language-server with the appropriate
nix-shell.

The haskell-language-server page has some instructions for integration
with various editors: https://github.com/haskell/haskell-language-server#editor-integration

In my case, using vscode, I needed to alter the workspace settings to
set the haskell-language-server binary to
${workspaceFolder}/scripts/haskell-language-server.