CAD-1861: Abstract over TxBody
This PR makes the transaction body abstract and prepares for overriding
it in subsequent eras.

- TxBody is now a type family, instantiated in Shelley to the existing
TxBody type.
- A new "package" type TxBodyConstraints is introduced, which collects a
bunch of the "utility" instances (NoThunks, Eq, CBOR etc) for TxBody.
- Where we have specific dependencies on parts of the TxBody, we use
'HasField' constraints. This will allow a future 'TxBody' to define
these fields and continue to work. The cost is that we must be much more
explicit in terms of our requirements.

  One may ask why these are not done as part of `TxBodyConstraints`.
  There are two reasons - one prosaic (it would cause import cycles) and
  one principled - we already know that some of these are going to need
  to change in Allegra (txfee becomes optional), so packaging them
  together requires us to re-define these in future.
- Various of the STS infrastructure is now specific to `ShelleyEra`,
because at the point of declaring the STS instances we must be
monomorphic. These will therefore need to be redeclared for future eras.
- As yet, we do not introduce lenses for state modifications, since we
have not abstracted over any state, merely part of the signal (i.e. the
TxBody). So we instead fix the state, environment etc of subsystems.
- At present the API is still general, so could be used for future eras.
- The tests are updated with appropriate constraints that should allow
them to continue to be used for any Shelley-like era.